Refine architecture: unified IDs, SessionId routing, sync epochs, async match delay
- ids.rs: Add Id<T>/IdGen<T> phantom-typed ID system; WorkerId/SessionId/PartyId/GroupId are now type-incompatible
- farm/supervisor.rs: Replace FarmModuleActor with FarmSupervisor; SessionRecord with slot mapping; direct SessionId routing without BiMap; lifecycle cleanup (on_worker_dead/on_session_dead); new Event variants (SessionReady/SessionDead/PartyRound/PartyDone)
- farm/party/actor.rs: Replace HashMap with IndexMap for deterministic master selection; versioned sync_epoch to eliminate clear_sync races; InMenuTask::total replaces end (fixes off-by-one); async match accept via tokio::spawn (non-blocking mailbox); Done event on engine stop; SessionId replaces MemberId throughout
- main.rs: Real handle_event with pending_sessions queue and PARTY_SIZE auto-group logic
- bin/simulate.rs: DSL-based SimSession harness for full-flow integration testing
- lib.rs: Export farm/ids modules as farm_core for binary reuse
- server stub: WorkerId changed from u32 to u64

+++ stubs/server/src/lib.rs
@@ -5,7 +5,7 @@ use kameo::actor::Recipient;
 use protocol::migo::{ClientType, worker};
 
 #[derive(Debug, Clone, Copy, Hash, Eq, PartialEq)]
-pub struct WorkerId(pub u32);
+pub struct WorkerId(pub u64);
 
 #[derive(Debug)]
 pub enum Event {
